+++ lib/cogment/utils.cpp
@@ -20,7 +20,7 @@
 
 namespace cogment {
 
-#ifdef __linux__
+#if defined(__linux__) || defined(__APPLE__)
   #include <string.h>
   #include <time.h>